% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BM_Density.R
\name{BM_Density}
\alias{BM_Density}
\title{Beta Mixture pdf}
\usage{
BM_Density(x, Mixture_Parameters, Alphas, Betas, Log = F)
}
\arguments{
\item{website}{the url of a player's page on eliteprospects.com.}
}
\value{
data frame (or list of data frames) with data from all the players drafted in the given year.
}
\description{
returns the density of a given point based on the given parameters of a beta mixture distribution
}
