% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BM_Fit.R
\name{BM_Fit}
\alias{BM_Fit}
\title{Beta Mixture Density Fitting Procedure}
\usage{
BM_Fit(x, K, threshold, seed = NA)
}
\arguments{
\item{x}{vector of data points. Will fit beta mixture density to this.}

\item{K}{Number of distributions composing mixture}

\item{threshold}{Determines "accuracy" of final returned parameters}

\item{seed}{Sets seed for random processes. Should only be used if replicable results are needed.}
}
\value{
TODO
}
\description{
Fits a Beta Mixture density to a given set of data
}
